created UIUtils to hold utility functions that help generate/manipulate UI elements. Moved in the most obvious functions I could think of, hope I forgot none :)
