+++ fe2-android/app/src/main/java/com/github/dedis/popstellar/ui/lao/InviteFragment.kt
@@ -12,7 +12,7 @@ import com.github.dedis.popstellar.model.Role
 import com.github.dedis.popstellar.model.qrcode.ConnectToLao
 import com.github.dedis.popstellar.repository.remote.GlobalNetworkManager
 import com.github.dedis.popstellar.utility.ActivityUtils.getQRCodeColor
-import com.github.dedis.popstellar.utility.GeneralUtils
+import com.github.dedis.popstellar.utility.UIUtils
 import com.github.dedis.popstellar.utility.error.ErrorUtils.logAndShow
 import com.github.dedis.popstellar.utility.error.UnknownLaoException
 import com.google.gson.Gson
@@ -27,7 +27,7 @@ class InviteFragment : Fragment() {
 
   private lateinit var laoViewModel: LaoViewModel
   private lateinit var binding: InviteFragmentBinding
-  private lateinit var clipboardManager: GeneralUtils.ClipboardUtil
+  private lateinit var clipboardManager: UIUtils.ClipboardUtil
 
   override fun onCreateView(
       inflater: LayoutInflater,
@@ -36,7 +36,7 @@ class InviteFragment : Fragment() {
   ): View? {
     binding = InviteFragmentBinding.inflate(inflater, container, false)
     laoViewModel = LaoActivity.obtainViewModel(requireActivity())
-    clipboardManager = GeneralUtils.ClipboardUtil(requireActivity())
+    clipboardManager = UIUtils.ClipboardUtil(requireActivity())
 
     // Display the LAO identifier, not the device public key
     binding.laoPropertiesIdentifierText.text = laoViewModel.laoId
